Added main class to scr/main/java
