+++ docs/CHANGELOG.md
@@ -20,7 +20,7 @@
 - [Noah Manning, Simeon Sodunke, Samuel Yohannes, Kavin Xayamongkhon, Melkamu Gonfa, Tariq Hassan] Uploaded UML Diagram 
 
  # Milestone 3
-- [Kavin X] Create Java file folder for MVC model, added ApplicationModel, ApplicationView, ApplicationController packages
-- [Kavin X] Update ApplicationModel/User Class, added methods and lists
-- [Kavin X] Update ApplicationView/User Class, added methods and lists
-- [Kavin X] Update ApplicationController/User Class, added methods and lists
+- [Kavin X] Create Java file folder for MVC model, added edu.metrostate.ApplicationModel, edu.metrostate.ApplicationView, edu.metrostate.ApplicationController packages
+- [Kavin X] Update edu.metrostate.ApplicationModel/User Class, added methods and lists
+- [Kavin X] Update edu.metrostate.ApplicationView/User Class, added methods and lists
+- [Kavin X] Update edu.metrostate.ApplicationController/User Class, added methods and lists
