--- src/main/java/ApplicationModel/Game.java
@@ -1,4 +1,0 @@
-package ApplicationModel;
-
-public class Game {
-}
