--- src/main/java/ApplicationModel/GameScores.java
@@ -1,4 +1,0 @@
-package ApplicationModel;
-
-public class GameScores {
-}
