--- src/main/java/ApplicationModel/LiveGame.java
@@ -1,4 +1,0 @@
-package ApplicationModel;
-
-public class LiveGame {
-}
