--- src/main/java/ApplicationModel/Player.java
@@ -1,5 +1,0 @@
-package ApplicationModel;
-
-public class Player {
-
-}
