--- src/main/java/ApplicationModel/Sport.java
@@ -1,4 +1,0 @@
-package ApplicationModel;
-
-public class Sport {
-}
