--- src/main/java/ApplicationModel/Ticket.java
@@ -1,4 +1,0 @@
-package ApplicationModel;
-
-public class Ticket {
-}
