+++ src/main/java/edu/metrostate/ApplicationModel/GameScores.java
@@ -1,0 +1,4 @@
+package edu.metrostate.ApplicationModel;
+
+public class GameScores {
+}
