+++ src/main/java/edu/metrostate/ApplicationModel/Player.java
@@ -1,0 +1,5 @@
+package edu.metrostate.ApplicationModel;
+
+public class Player {
+
+}
