+++ src/main/java/main/Main.java
@@ -1,9 +1,5 @@
 package main;
 
-import ApplicationController.UserController;
-import ApplicationModel.User;
-import ApplicationView.UserView;
-
 public class Main {
     public static void main(String[] args) {
         System.out.println("test");
